backend/requirements.txt
fastapi==0.109.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.0
python-dotenv==1.0.0
